backend/news_handler/requirements.txt
newspaper4k
lxml[html_clean]
requests
python-dotenv
sentence-transformers
openai
pymongo[srv]